==> notebooks/14_inference_pipeline.ipynb <==
%load_ext autoreload
%autoreload 2

import os
os.chdir('..')

from datetime import datetime, timedelta
import pandas as pd

current_date = pd.to_datetime(datetime.now()).floor('H')
print(f'{current_date=}')
# current_date = pd.Timestamp('2023-02-28 09:00:00')

from src.inference import load_batch_of_features_from_store

features = load_batch_of_features_from_store(current_date)

print(features)

from src.inference import (
    load_model_from_registry,
    get_model_predictions
)


#predictions = get_model_predictions(model, features)

model = load_model_from_registry()

predictions = get_model_predictions(model, features)

print(predictions)

predictions['pickup_hour'] = current_date
predictions

from src.inference import get_feature_store
import src.config as config

# connect to the feature group
feature_group = get_feature_store().get_or_create_feature_group(
    name=config.FEATURE_GROUP_MODEL_PREDICTIONS,
    version=1,
    description="Predictions generate by our production model",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour',
)

feature_group.insert(predictions, write_options={"wait_for_job": False})

import hopsworks

# connect to the project
project = hopsworks.login(
    project=config.HOPSWORKS_PROJECT_NAME,
    api_key_value=config.HOPSWORKS_API_KEY
)

# connect to the feature store
feature_store1 = project.get_feature_store()

# connect to the feature group
feature_group1 = feature_store1.get_feature_group(
    name='model_predictions_feature_group',
    version=1,
)

# create feature view (if it doesn't exist yet)
# This feature view only uses on feature group, so the query is trivial

# create feature view if it doesn't exist yet
feature_store1.create_feature_view(
        name=config.FEATURE_VIEW_PREDICTIONS_NAME,
        version=1,
        query=feature_group1.select_all()
    )



import src
project=src.inference.get_hopsworks_project()
store=project.get_feature_store()
feature_group=store.get_feature_group(
        name=config.FEATURE_GROUP_MODEL_PREDICTIONS,version=1
    )
results=store.get_feature_view(config.FEATURE_VIEW_PREDICTIONS_NAME,version=1)

p = results.get_batch_data(start_time=current_date-timedelta(days=28),end_time=current_date)

print(p)
